+++ lib/middleware-common/src/block_trace.rs
@@ -25,6 +25,7 @@ impl FunctionMiddleware for BlockTrace {
         op: Event<'a, 'b>,
         _module_info: &ModuleInfo,
         sink: &mut EventSink<'a, 'b>,
+        _source_loc: u32,
     ) -> Result<(), Self::Error> {
         match op {
             Event::Internal(InternalEvent::FunctionBegin(_)) => {