+++ lib/runtime-core/src/lib.rs
@@ -66,6 +66,8 @@ pub mod vmcalls;
 pub use trampoline_x64 as trampoline;
 #[cfg(unix)]
 pub mod fault;
+#[cfg(feature = "generate-debug-information")]
+pub mod jit_debug;
 pub mod state;
 #[cfg(feature = "managed")]
 pub mod tiering;